+++ codex-ui/dev/Playground.vue
@@ -27,208 +27,15 @@
         <router-view />
       </div>
     </div>
-    <Heading :level="3">
-      Avatar
-    </Heading>
-    <Avatar
-      src="../static/example-avatar.png"
-      username="Vitaly"
-    />
-
-    <Heading :level="3">
-      Row
-    </Heading>
-
-    <Row
-      title="Title"
-      subtle="This item is no longer detected near you. It was last seen near Pesochnaya Embankment, 14 литТ к2."
-      :has-delimiter="true"
-    >
-      <template #left>
-        <Avatar
-          src="../static/example-avatar.png"
-          username="Vitaly"
-        />
-      </template>
-
-      <template #right>
-        <Button
-          size="small"
-          :style="'secondary'"
-        >
-          Edit
-        </Button>
-      </template>
-    </Row>
-
-    <Row
-      title="Title"
-      subtle="This item is no longer detected near you. It was last seen near Pesochnaya Embankment, 14 литТ к2."
-      label="Time Sensitive"
-    >
-      <template #left>
-        <Avatar
-          src="../static/example-avatar.png"
-          username="Vitaly"
-        />
-      </template>
-
-      <template #right>
-        <Button
-          size="small"
-          :style="'secondary'"
-        >
-          Edit
-        </Button>
-      </template>
-    </Row>
-
-    <Heading :level="3">
-      Icon
-    </Heading>
-
-    <Icon name="Checklist" /> <Icon name="Loader" />
-    <br>
-    <Heading :level="3">
-      Switch
-    </Heading>
-    on, default
-    <Switch value />
-    <br>
-    on, disabled
-    <Switch
-      value
-      disabled
-    />
-    <br>
-    off, default
-    <Switch />
-    <br>
-    off, disabled
-    <Switch disabled />
-    <br>
-    <br>
-    <Heading :level="3">
-      Form Section
-    </Heading>
-
-    <Section
-      title="List Name"
-      caption="Item list"
-    >
-      <Row
-        v-for="(item, index) in formSectionItems"
-        :key="item.id"
-        :title="item.name"
-        :has-delimiter="index !== formSectionItems.length - 1"
-      >
-        <template #left>
-          <Avatar
-            src="../static/example-avatar.png"
-            :username="item.name"
-          />
-        </template>
-
-        <template #right>
-          <Button
-            size="small"
-            :style="'secondary'"
-          >
-            Can View
-          </Button>
-        </template>
-      </Row>
-    </Section>
-    <br>
-    <Heading :level="3">
-      Tab
-    </Heading>
-    <Tab
-      id="/"
-      title="Home"
-    />
-    <br>
-    <Tab
-      id="/"
-      title="Home"
-      picture="../static/example-avatar.png"
-    />
-    <br>
-    <Tab
-      id="/"
-      title="Home"
-      icon="Plus"
-    />
-    <br>
-    <Tab
-      id="/"
-      title="Home"
-      :closable="true"
-    />
-    <br>
-    <Tab
-      id="/"
-      title="Home"
-      :closable="true"
-      :is-active="true"
-    />
-    <Heading :level="3">
-      Tabbar
-    </Heading>
-    <Heading :level="4">
-      Tabbar default
-    </Heading>
-    <Tabbar
-      :tabs="defaultTabs"
-    />
-    <Heading :level="4">
-      Tabbar without crosses
-    </Heading>
-    <Tabbar
-      :tabs="tabsWithoutCrossState"
-    />
-    <Heading :level="3">
-      Context Menu
-    </Heading>
-    <ContextMenu
-      :show-search="true"
-      :items="contextMenuItems"
-    />
-    <Heading :level="3">
-      Vertical Menu
-    </Heading>
-    <VerticalMenu :items="verticalMenuItems" />
-    <Heading :level="3">
-      Type Scale
-    </Heading>
-    <TypeScale />
-    <Heading :level="3">
-      Editor.js
-    </Heading>
-    <Editor :tools="{}" />
   </div>
 </template>
 
 <script setup lang="ts">
 import { computed } from 'vue';
 import {
-  Button,
-  Heading,
-  Editor,
-  Section,
-  Icon,
-  Row,
-  Avatar,
-  Switch,
-  Tab,
-  ContextMenu,
-  ContextMenuItem,
   VerticalMenu,
-  VerticalMenuItem,
-  Tabbar,
-  TabParams
+  Tabbar
 } from '../src/vue';
-import TypeScale from './TypeScale.vue';
 
 import { useRouter, useRoute } from 'vue-router';
 
@@ -237,6 +44,8 @@ const route = useRoute();
 const pages = computed(() => [
   {
     title: 'TypeScale',
+    onActivate: () => router.push('/type-scale'),
+    isActive: route.path === '/type-scale',
   },
   {
     title: 'Theming',
@@ -344,127 +153,6 @@ const pages = computed(() => [
     ],
   },
 ]);
-
-/**
- * Form section items elements
- */
-const formSectionItems = [
-  { id: 1,
-    name: 'Vitaly' },
-  { id: 2,
-    name: 'Nickmel' },
-];
-
-/**
- * Items for searching in the context menu
- */
-const contextMenuItems: ContextMenuItem[] = [
-  {
-    type: 'default',
-    title: 'Header 1',
-    icon: 'H1',
-    onActivate: doNothing,
-  },
-  {
-    title: 'Header 2',
-    icon: 'H1',
-    onActivate: doNothing,
-  },
-  {
-    type: 'separator',
-  },
-  {
-    type: 'default',
-    title: 'Header 3',
-    icon: 'H1',
-    onActivate: doNothing,
-  },
-  {
-    type: 'default',
-    title: 'Header 3',
-    onActivate: doNothing,
-  },
-];
-
-/**
- * Tabs for example usage of tabbar
- */
-const defaultTabs = [
-  {
-    title: 'Home',
-    id: '/',
-    picture: '../static/example-avatar.png',
-  },
-  {
-    title: 'New note',
-    id: '/new',
-    closable: true,
-  },
-  {
-    title: 'User settings',
-    id: '/settings',
-    closable: true,
-    isActive: true,
-  },
-];
-
-const tabsWithoutCrossState: TabParams[] = [
-  {
-    title: 'Home',
-    id: '/',
-  },
-  {
-    title: 'New note',
-    id: '/',
-    closable: true,
-  },
-  {
-    title: 'User settings',
-    id: '/',
-    closable: true,
-    isActive: true,
-  },
-];
-
-for (let _ = 0; _ < 15; ++_) {
-  tabsWithoutCrossState.push({
-    title: 'Marketplace',
-    id: '/marketplace',
-    closable: true,
-    icon: 'Plus',
-  });
-}
-
-/**
- * Simple function for example
- */
-function doNothing(): void {}
-
-/**
- * Items for displaing in vertical menu
- */
-const verticalMenuItems: VerticalMenuItem[] = [
-  {
-    title: 'CodeX',
-    items: [
-      {
-        title: 'Workflow',
-      },
-      {
-        title: 'Enineering Strategy',
-        items: [
-          {
-            title: 'SSH Keys',
-            isActive: true,
-          },
-        ],
-      },
-      {
-        title: 'Admission Campaign 2022',
-      },
-    ],
-  },
-];
 </script>
 
 <style module>
