+++ codex-ui/dev/pages/components/Editor.vue
@@ -5,10 +5,13 @@
       Component for creating a rich text content using Editor.js
     </template>
   </PageHeader>
+
+  <Editor :tools="{}" />
 </template>
 
 <script setup lang="ts">
 import PageHeader from '../../components/PageHeader.vue';
+import { Editor } from '../../../src/vue';
 
 </script>
