+++ codex-ui/dev/pages/components/Switch.vue
@@ -5,6 +5,23 @@
       Best used for activating or deactivating features or modes within an application. The change is usually effective immediately.
     </template>
   </PageHeader>
+
+  <div class="showcase">
+    on, default
+    <Switch value />
+    <br>
+    on, disabled
+    <Switch
+      value
+      disabled
+    />
+    <br>
+    off, default
+    <Switch />
+    <br>
+    off, disabled
+    <Switch disabled />
+  </div>
 </template>
 
 <script setup lang="ts">
@@ -14,4 +31,9 @@ import { Switch } from '../../../src/vue';
 </script>
 
 <style scoped>
+.showcase {
+  display: grid;
+  gap: var(--spacing-l);
+  grid-template-columns: repeat(1, max-content);
+}
 </style>
